+++ contracts/test/TierReportTest.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: CAL
 pragma solidity 0.8.10;
 
-import { ITier } from "../tier/ITier.sol";
-import { TierReport } from "../tier/libraries/TierReport.sol";
+import {ITier} from "../tier/ITier.sol";
+import {TierReport} from "../tier/libraries/TierReport.sol";
 
 /// @title TierReportTest
 /// Thin wrapper around the `TierReport` library for hardhat unit testing.
@@ -10,10 +10,10 @@ contract TierReportTest {
     /// Wraps `TierReport.tierAtBlockFromReport`.
     /// @param report_ Forwarded to TierReport.
     /// @param blockNumber_ Forwarded to TierReport.
-    function tierAtBlockFromReport(uint report_, uint blockNumber_)
+    function tierAtBlockFromReport(uint256 report_, uint256 blockNumber_)
         external
         pure
-        returns (uint)
+        returns (uint256)
     {
         return TierReport.tierAtBlockFromReport(report_, blockNumber_);
     }
@@ -21,10 +21,10 @@ contract TierReportTest {
     /// Wraps `TierReport.tierBlock`.
     /// @param report_ Forwarded to TierReport.
     /// @param tier_ Forwarded to TierReport.
-    function tierBlock(uint report_, uint tier_)
+    function tierBlock(uint256 report_, uint256 tier_)
         external
         pure
-        returns (uint)
+        returns (uint256)
     {
         return TierReport.tierBlock(report_, tier_);
     }
@@ -32,10 +32,10 @@ contract TierReportTest {
     /// Wraps `TierReport.truncateTiersAbove`.
     /// @param report_ Forwarded to TierReport.
     /// @param tier_ Forwarded to TierReport.
-    function truncateTiersAbove(uint report_, uint tier_)
+    function truncateTiersAbove(uint256 report_, uint256 tier_)
         external
         pure
-        returns (uint)
+        returns (uint256)
     {
         return TierReport.truncateTiersAbove(report_, tier_);
     }
@@ -46,11 +46,11 @@ contract TierReportTest {
     /// @param endTier_ Forwarded to TestUtil.
     /// @param blockNumber_ Forwarded to TestUtil.
     function updateBlocksForTierRange(
-        uint report_,
-        uint startTier_,
-        uint endTier_,
-        uint blockNumber_
-    ) external pure returns (uint) {
+        uint256 report_,
+        uint256 startTier_,
+        uint256 endTier_,
+        uint256 blockNumber_
+    ) external pure returns (uint256) {
         return
             TierReport.updateBlocksForTierRange(
                 report_,
@@ -66,11 +66,11 @@ contract TierReportTest {
     /// @param endTier_ Forwarded to TestUtil.
     /// @param blockNumber_ Forwarded to TestUtil.
     function updateReportWithTierAtBlock(
-        uint report_,
-        uint startTier_,
-        uint endTier_,
-        uint blockNumber_
-    ) external pure returns (uint) {
+        uint256 report_,
+        uint256 startTier_,
+        uint256 endTier_,
+        uint256 blockNumber_
+    ) external pure returns (uint256) {
         return
             TierReport.updateReportWithTierAtBlock(
                 report_,
